move struct to util.go
